refactor: complete medium boot tree migration into core/boot
