+++ CMakeLists.txt
@@ -339,6 +339,7 @@ endif()
 # Add subdirectories
 
 add_subdirectory(arch)
+add_subdirectory(core/boot)
 add_subdirectory(boot)
 add_subdirectory(hal)
 add_subdirectory(platform)
